I refactored the form display process to model the form into a display agnostic data structure, then allow that structure to build the Java display process.
Checks, Fields, Labels, Lines and Pages work.  
There is something not functioning in displaying comboboxes that seems to get them mixed up with other components, I don't have time to figure it out right now.  

Still need to implment cells which will require the ability to reference other pages, which is what started the refactoring.  In order to do this we pages will need to have awareness of the document so that they can ask for information that exists in other pages.