Comboboxes now display!
Program is choking on 638s, actually it is choking at an element named "CHECK2" on the 638, I think the problem is with the fact that this particular element has an incomplete (or abnormal) <itemlocation> definition.  I'll need to reread the standard and determine what to do when only the location and not location and size are listed.  Regardless it results in a null return when trying to process the value, which shouldn't happen. Situation needs review.